Setup the first test with jest. Some refactoring to move files under the proper folders
