+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test": "jest"
   },
   "dependencies": {
     "@emotion/react": "^11.11.4",
@@ -30,6 +31,9 @@
     "typescript": "4.9.3"
   },
   "devDependencies": {
-    "@types/styled-components": "^5.1.34"
+    "@types/jest": "^29.5.14",
+    "@types/styled-components": "^5.1.34",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.5"
   }
 }
