+++ pages/movieDetails/[id].tsx
@@ -5,7 +5,7 @@ import {GetServerSideProps} from "next";
 import {Movie, MovieDetail} from "../../data/movies/Movie";
 import {MovieDetailCard} from "../../components/MovieDetailCard";
 import {FloatingBackButton} from "../../components/FloatingBackButton";
-import {getMovieDetailsUseCase} from "../../utils/MovieDetailUseCase";
+import {getMovieDetailsUseCase} from "../../src/MovieDetailUseCase";
 
 const Wrapper = styled.div`
     display: flex;
